backend/src/requirements.txt
transformers>=4.40.0
accelerate
sentencepiece
pillow
pymupdf
fastapi
uvicorn
python-multipart
sqlalchemy
redis
rq
prometheus-client
psycopg2-binary
python-dotenv